+++ src/WeihanLi.Common/Data/Repository.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Linq.Expressions;
 using System.Reflection;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using WeihanLi.Common.Models;
 using WeihanLi.Extensions;
@@ -51,7 +52,7 @@ SELECT COUNT(1) FROM {TableName}
             return _dbConnection.Value.ExecuteScalarTo<int>(sql, whereSql.Parameters);
         }
 
-        public virtual Task<int> CountAsync(Expression<Func<TEntity, bool>> whereExpression)
+        public virtual Task<int> CountAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
 
@@ -59,7 +60,7 @@ SELECT COUNT(1) FROM {TableName}
 SELECT COUNT(1) FROM {TableName}
 {whereSql.SqlText}
 ";
-            return _dbConnection.Value.ExecuteScalarToAsync<int>(sql, whereSql.Parameters);
+            return _dbConnection.Value.ExecuteScalarToAsync<int>(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual long LongCount(Expression<Func<TEntity, bool>> whereExpression)
@@ -73,7 +74,7 @@ SELECT COUNT(1) FROM {TableName}
             return _dbConnection.Value.ExecuteScalarTo<long>(sql, whereSql.Parameters);
         }
 
-        public virtual Task<long> LongCountAsync(Expression<Func<TEntity, bool>> whereExpression)
+        public virtual Task<long> LongCountAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
 
@@ -81,7 +82,7 @@ SELECT COUNT(1) FROM {TableName}
 SELECT COUNT(1) FROM {TableName}
 {whereSql.SqlText}
 ";
-            return _dbConnection.Value.ExecuteScalarToAsync<long>(sql, whereSql.Parameters);
+            return _dbConnection.Value.ExecuteScalarToAsync<long>(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual bool Exist(Expression<Func<TEntity, bool>> whereExpression)
@@ -91,11 +92,11 @@ SELECT COUNT(1) FROM {TableName}
             return _dbConnection.Value.ExecuteScalarTo<bool>(sql, whereSql.Parameters);
         }
 
-        public virtual Task<bool> ExistAsync(Expression<Func<TEntity, bool>> whereExpression)
+        public virtual Task<bool> ExistAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var sql = $@"SELECT CAST(IIF(EXISTS (SELECT TOP(1) 1 FROM {TableName} {whereSql.SqlText}), 1, 0) AS BIT)";
-            return _dbConnection.Value.ExecuteScalarToAsync<bool>(sql, whereSql.Parameters);
+            return _dbConnection.Value.ExecuteScalarToAsync<bool>(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual TEntity Fetch(Expression<Func<TEntity, bool>> whereExpression)
@@ -108,14 +109,14 @@ SELECT TOP(1) {SelectColumnsString} FROM {TableName}
             return _dbConnection.Value.Fetch<TEntity>(sql, whereSql.Parameters);
         }
 
-        public virtual Task<TEntity> FetchAsync(Expression<Func<TEntity, bool>> whereExpression)
+        public virtual Task<TEntity> FetchAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var sql = $@"
 SELECT TOP 1 {SelectColumnsString} FROM {TableName}
 {whereSql.SqlText}
 ";
-            return _dbConnection.Value.FetchAsync<TEntity>(sql, whereSql.Parameters);
+            return _dbConnection.Value.FetchAsync<TEntity>(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual TEntity Fetch<TProperty>(Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
@@ -129,7 +130,7 @@ ORDER BY {GetColumnName(orderByExpression.GetMemberName())}  {(ascending ? "" : 
             return _dbConnection.Value.Fetch<TEntity>(sql, whereSql.Parameters);
         }
 
-        public virtual Task<TEntity> FetchAsync<TProperty>(Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
+        public virtual Task<TEntity> FetchAsync<TProperty>(Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var sql = $@"
@@ -137,7 +138,7 @@ SELECT TOP(1) {SelectColumnsString} FROM {TableName}
 {whereSql.SqlText}
 ORDER BY {GetColumnName(orderByExpression.GetMemberName())}  {(ascending ? "" : "DESC")}
 ";
-            return _dbConnection.Value.FetchAsync<TEntity>(sql, whereSql.Parameters);
+            return _dbConnection.Value.FetchAsync<TEntity>(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual List<TEntity> Select(Expression<Func<TEntity, bool>> whereExpression)
@@ -150,14 +151,14 @@ SELECT {SelectColumnsString} FROM {TableName}
             return _dbConnection.Value.Select<TEntity>(sql, whereSql.Parameters).ToList();
         }
 
-        public virtual Task<List<TEntity>> SelectAsync(Expression<Func<TEntity, bool>> whereExpression)
+        public virtual Task<List<TEntity>> SelectAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var sql = $@"
 SELECT {SelectColumnsString} FROM {TableName}
 {whereSql.SqlText}
 ";
-            return _dbConnection.Value.SelectAsync<TEntity>(sql, whereSql.Parameters).ContinueWith(r => r.Result.ToList());
+            return _dbConnection.Value.SelectAsync<TEntity>(sql, whereSql.Parameters, cancellationToken: cancellationToken).ContinueWith(r => r.Result.ToList(), cancellationToken);
         }
 
         public virtual List<TEntity> Select<TProperty>(int count, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
@@ -171,7 +172,7 @@ ORDER BY {GetColumnName(orderByExpression.GetMemberName())} {(ascending ? "" : "
             return _dbConnection.Value.Select<TEntity>(sql, whereSql.Parameters).ToList();
         }
 
-        public virtual Task<List<TEntity>> SelectAsync<TProperty>(int count, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
+        public virtual Task<List<TEntity>> SelectAsync<TProperty>(int count, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var sql = $@"
@@ -179,10 +180,10 @@ SELECT TOP({count}) {SelectColumnsString} FROM {TableName}
 {whereSql.SqlText}
 ORDER BY {GetColumnName(orderByExpression.GetMemberName())} {(ascending ? "" : "DESC")}
 ";
-            return _dbConnection.Value.SelectAsync<TEntity>(sql, whereSql.Parameters).ContinueWith(_ => _.Result.ToList());
+            return _dbConnection.Value.SelectAsync<TEntity>(sql, whereSql.Parameters, cancellationToken: cancellationToken).ContinueWith(_ => _.Result.ToList(), cancellationToken);
         }
 
-        public virtual PagedListModel<TEntity> Paged<TProperty>(int pageNumber, int pageSize, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
+        public virtual IPagedListModel<TEntity> Paged<TProperty>(int pageNumber, int pageSize, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             if (pageNumber <= 0)
@@ -216,7 +217,7 @@ FETCH NEXT {pageSize} ROWS ONLY
             return _dbConnection.Value.Select<TEntity>(sql, whereSql.Parameters).ToPagedListModel(pageNumber, pageSize, total);
         }
 
-        public virtual async Task<PagedListModel<TEntity>> PagedAsync<TProperty>(int pageNumber, int pageSize, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false)
+        public virtual async Task<IPagedListModel<TEntity>> PagedAsync<TProperty>(int pageNumber, int pageSize, Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> orderByExpression, bool ascending = false, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             if (pageNumber <= 0)
@@ -231,7 +232,7 @@ FETCH NEXT {pageSize} ROWS ONLY
 SELECT COUNT(1) FROM {TableName}
 {whereSql.SqlText}
 ";
-            var total = await _dbConnection.Value.ExecuteScalarToAsync<int>(sql, whereSql.Parameters);
+            var total = await _dbConnection.Value.ExecuteScalarToAsync<int>(sql, whereSql.Parameters, cancellationToken: cancellationToken);
             if (total == 0)
             {
                 return new TEntity[0].ToPagedListModel(pageNumber, pageSize, 0);
@@ -247,7 +248,7 @@ OFFSET {offset} ROWS
 FETCH NEXT {pageSize} ROWS ONLY
 ";
 
-            return (await _dbConnection.Value.SelectAsync<TEntity>(sql, whereSql.Parameters)).ToPagedListModel(pageNumber, pageSize, total);
+            return (await _dbConnection.Value.SelectAsync<TEntity>(sql, whereSql.Parameters, cancellationToken: cancellationToken)).ToPagedListModel(pageNumber, pageSize, total);
         }
 
         public virtual int Insert(TEntity entity)
@@ -271,7 +272,7 @@ FETCH NEXT {pageSize} ROWS ONLY
             return _dbConnection.Value.Execute(sql, paramDictionary);
         }
 
-        public virtual Task<int> InsertAsync(TEntity entity)
+        public virtual Task<int> InsertAsync(TEntity entity, CancellationToken cancellationToken = default)
         {
             var paramDictionary = new Dictionary<string, object>();
 
@@ -287,7 +288,7 @@ FETCH NEXT {pageSize} ROWS ONLY
             }
 
             sqlBuilder.AppendLine(")");
-            return _dbConnection.Value.ExecuteAsync(sqlBuilder.ToString(), paramDictionary);
+            return _dbConnection.Value.ExecuteAsync(sqlBuilder.ToString(), paramDictionary, cancellationToken: cancellationToken);
         }
 
         public virtual int Insert(IEnumerable<TEntity> entities)
@@ -324,7 +325,7 @@ FETCH NEXT {pageSize} ROWS ONLY
             return _dbConnection.Value.Execute(sqlBuilder.ToString(), paramDictionary);
         }
 
-        public virtual Task<int> InsertAsync(IEnumerable<TEntity> entities)
+        public virtual Task<int> InsertAsync(IEnumerable<TEntity> entities, CancellationToken cancellationToken = default)
         {
             var count = entities?.Count() ?? 0;
             if (count == 0)
@@ -355,7 +356,7 @@ FETCH NEXT {pageSize} ROWS ONLY
             }
             sqlBuilder.Remove(sqlBuilder.Length - 2, 1);
 
-            return _dbConnection.Value.ExecuteAsync(sqlBuilder.ToString(), paramDictionary);
+            return _dbConnection.Value.ExecuteAsync(sqlBuilder.ToString(), paramDictionary, cancellationToken: cancellationToken);
         }
 
         public virtual int Update<TProperty>(Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> propertyExpression, object value)
@@ -371,7 +372,7 @@ SET {GetColumnName(propertyName)} = @set_{propertyName}
             return _dbConnection.Value.Execute(sql, whereSql.Parameters);
         }
 
-        public virtual Task<int> UpdateAsync<TProperty>(Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> propertyExpression, object value)
+        public virtual Task<int> UpdateAsync<TProperty>(Expression<Func<TEntity, bool>> whereExpression, Expression<Func<TEntity, TProperty>> propertyExpression, object value, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var propertyName = propertyExpression.GetMemberName();
@@ -381,7 +382,7 @@ SET {GetColumnName(propertyName)} = @set_{propertyName}
 {whereSql.SqlText}
 ";
             whereSql.Parameters.Add($"set_{propertyName}", value);
-            return _dbConnection.Value.ExecuteAsync(sql, whereSql.Parameters);
+            return _dbConnection.Value.ExecuteAsync(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual int Update(Expression<Func<TEntity, bool>> whereExpression, IDictionary<string, object> propertyValues)
@@ -403,7 +404,7 @@ SET {propertyValues.Keys.Select(p => $"{GetColumnName(p)}=@set_{p}").StringJoin(
             return _dbConnection.Value.Execute(sql, whereSql.Parameters);
         }
 
-        public virtual Task<int> UpdateAsync(Expression<Func<TEntity, bool>> whereExpression, IDictionary<string, object> propertyValues)
+        public virtual Task<int> UpdateAsync(Expression<Func<TEntity, bool>> whereExpression, IDictionary<string, object> propertyValues, CancellationToken cancellationToken = default)
         {
             if (propertyValues == null || propertyValues.Count == 0)
             {
@@ -419,7 +420,7 @@ SET {propertyValues.Keys.Select(p => $"{GetColumnName(p)}=@set_{p}").StringJoin(
             {
                 whereSql.Parameters.Add($"set_{propertyValue.Key}", propertyValue.Value);
             }
-            return _dbConnection.Value.ExecuteAsync(sql, whereSql.Parameters);
+            return _dbConnection.Value.ExecuteAsync(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual int Delete(Expression<Func<TEntity, bool>> whereExpression)
@@ -432,29 +433,29 @@ DELETE FROM {TableName}
             return _dbConnection.Value.Execute(sql, whereSql.Parameters);
         }
 
-        public virtual Task<int> DeleteAsync(Expression<Func<TEntity, bool>> whereExpression)
+        public virtual Task<int> DeleteAsync(Expression<Func<TEntity, bool>> whereExpression, CancellationToken cancellationToken = default)
         {
             var whereSql = SqlExpressionParser.ParseWhereExpression(whereExpression, ColumnMappings);
             var sql = $@"
 DELETE FROM {TableName}
 {whereSql.SqlText}
 ";
-            return _dbConnection.Value.ExecuteAsync(sql, whereSql.Parameters);
+            return _dbConnection.Value.ExecuteAsync(sql, whereSql.Parameters, cancellationToken: cancellationToken);
         }
 
         public virtual int Execute(string sqlStr, object param = null)
         => _dbConnection.Value.Execute(sqlStr, paramInfo: param);
 
-        public virtual Task<int> ExecuteAsync(string sqlStr, object param = null)
-        => _dbConnection.Value.ExecuteAsync(sqlStr, paramInfo: param);
+        public virtual Task<int> ExecuteAsync(string sqlStr, object param = null, CancellationToken cancellationToken = default)
+        => _dbConnection.Value.ExecuteAsync(sqlStr, paramInfo: param, cancellationToken: cancellationToken);
 
         public virtual TResult ExecuteScalar<TResult>(string sqlStr, object param = null)
 
         => _dbConnection.Value.ExecuteScalarTo<TResult>(sqlStr, paramInfo: param);
 
-        public virtual Task<TResult> ExecuteScalarAsync<TResult>(string sqlStr, object param = null)
+        public virtual Task<TResult> ExecuteScalarAsync<TResult>(string sqlStr, object param = null, CancellationToken cancellationToken = default)
 
-        => _dbConnection.Value.ExecuteScalarToAsync<TResult>(sqlStr, paramInfo: param);
+        => _dbConnection.Value.ExecuteScalarToAsync<TResult>(sqlStr, paramInfo: param, cancellationToken: cancellationToken);
 
         private static string GetColumnName(string propertyName)
         {